Remove text from footer
